Web: move interop_web to .js file instead of continuously trying to make emcc (and myself) despair by embedding large amounts of JavaScript in EM_ASM
